# packages/server/docs/swagger/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Mail Magic API",
		"version": "1.0.42",
		"description": "OpenAPI definition for the Mail Magic server. Authenticated endpoints require an API key provided as `Authorization: Bearer apikey-<token>`."
	},
	"servers": [
		{
			"url": "http://localhost:3776",
			"description": "Local development server"
		}
	],
...
	"paths": {
...
		"/api/v1/assets": {
			"post": {
				"tags": ["assets"],
				"summary": "Upload assets",
				"description": "Auth: API key. Upload one or more files to the domain asset directory or to a specific template directory. Requires multipart/form-data.",
				"security": [
					{
						"apiKeyBearer": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/AssetsUploadMultipartRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Assets uploaded.",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										},
										{
											"type": "object",
											"properties": {
												"data": {
													"$ref": "#/components/schemas/StatusOkData"
												}
											},
											"required": ["data"]
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Bad request.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"404": {
						"description": "Template not found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"413": {
						"description": "Upload too large.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				}
			}
		},
		"/asset/{domain}/{path}": {
			"get": {
				"tags": ["public-assets"],
				"summary": "Fetch a public asset",
				"description": "Public route for domain assets. The `{path}` parameter can contain slashes.",
				"security": [],
				"parameters": [
					{
						"name": "domain",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						},
						"description": "Domain name."
					},
					{
						"name": "path",
						"in": "path",
						"required": true,
						"allowReserved": true,
						"schema": {
							"type": "string"
						},
						"description": "Asset path under the domain assets directory (may contain `/`)."
					}
				],
				"responses": {
					"200": {
						"description": "Asset bytes.",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"404": {
						"description": "Not found."
					},
					"500": {
						"description": "Server error."
					}
				}
			}
		},
		"/api/asset/{domain}/{path}": {
			"get": {
				"tags": ["public-assets"],
				"summary": "Fetch a public asset (under API base path)",
				"description": "Compatibility route for assets reachable under API_BASE_PATH + ASSET_ROUTE.",
				"security": [],
				"parameters": [
					{
						"name": "domain",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "path",
						"in": "path",
						"required": true,
						"allowReserved": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Asset bytes.",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"404": {
						"description": "Not found."
					},
					"500": {
						"description": "Server error."
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"apiKeyBearer": {
				"type": "http",
				"scheme": "bearer",
				"description": "API key via Authorization: Bearer apikey-<secret>."
			}
		},
		"schemas": {
			"ApiResponse": {
				"type": "object",
				"description": "ApiResponse<T> standard envelope used by API endpoints. T is carried in the data field.",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"code": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"nullable": true
					},
					"errors": {
						"type": "object",
						"additionalProperties": {
							"type": "string"
						}
					}
				},
				"required": ["success", "code", "message", "data", "errors"]
			},
...
			"StatusOkData": {
				"type": "object",
				"properties": {
					"Status": {
						"type": "string",
						"examples": ["OK"]
					}
				},
				"required": ["Status"]
			},
...
			"AssetsUploadMultipartRequest": {
				"type": "object",
				"properties": {
					"domain": {
						"type": "string"
					},
					"path": {
						"type": "string",
						"description": "Optional subdirectory under assets root."
					},
					"dir": {
						"type": "string",
						"description": "Alias for path."
					},
					"templateType": {
						"type": "string",
						"description": "If set, upload to a template directory. Valid values: tx, form."
					},
					"template_type": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"template": {
						"type": "string",
						"description": "Template name/id (transactional) or form idname (form)."
					},
					"name": {
						"type": "string"
					},
					"idname": {
						"type": "string"
					},
					"formid": {
						"type": "string"
					},
					"locale": {
						"type": "string"
					},
					"files": {
						"type": "array",
						"items": {
							"type": "string",
							"format": "binary"
						},
						"minItems": 1
					}
				},
				"required": ["domain", "files"]
			}
		}
	}
}